% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_FR713_714_tables.R
\name{generate_FR713_714_tables}
\alias{generate_FR713_714_tables}
\title{generate_FR713_714_tables}
\usage{
generate_FR713_714_tables(
  fn011,
  fn022,
  fn023,
  fn024,
  fn025,
  fn026,
  fn028,
  fn111,
  fn112,
  fn121,
  fn123
)
}
\arguments{
\item{fn011}{FN project definition table FN011 which lists project dates, year, lake name and project leader information}

\item{fn022}{FN strata definition file for SSN}

\item{fn023}{FN strata definition file for DTP i.e. workday and weekend}

\item{fn024}{FN stratum definition for period i.e. am or pm}

\item{fn025}{FN stratum definition listing the exception days i.e. holidays that need to be re-coded as weekends}

\item{fn026}{FN stratum definition for the SPACE}

\item{fn028}{FN stratum definition for the MODE}

\item{fn111}{FN portion of the CLF that outlines the SAMA, Date, DOW, STRATUM,}

\item{fn112}{FN portion of the CLF that outlines the activity count, time of the count and number of interviews that occurred. Note: multiple counts could occur per SAMA}

\item{fn121}{FN CIF creel interview form}

\item{fn123}{FN species records part of the CIF form but nested by SAM to the FN121}
}
\value{
FN2-like output of creel results
}
\description{
This function generates the estimates and outputs a table for EFFORT and HARVEST
}
\examples{
\dontrun{
generate_FR713_714_tables(
SC00$FN011, SC00$FN022, SC00$FN023, SC00$FN024,
SC00$FN025, SC00$FN026, SC00$FN028, SC00$FN111,
SC00$FN112, SC00$FN121, SC00$FN123
)
}

}
