% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_SSN_LENGTH.R
\name{generate_SSN_LENGTH}
\alias{generate_SSN_LENGTH}
\title{Generate Season Length}
\usage{
generate_SSN_LENGTH(fn022, fn023, fn024, fn025)
}
\arguments{
\item{fn022}{FR022 FN2 design table}

\item{fn023}{FR023 FN2 design table}

\item{fn024}{FR023 FN2 design table}

\item{fn025}{FR023 FN2 design table}
}
\value{
summary table of SSN length
}
\description{
This function generates the season length give the FN2 design tables
correcting for exception days from FN025
}
