% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_creesys.R
\name{hello_creesys}
\alias{hello_creesys}
\title{hello creesys}
\usage{
hello_creesys()
}
\value{
str "hello creesys"
}
\description{
This function only returns 'hello creesys' but
is here as a simple example of how to write functions.
}
\examples{
hello_creesys()
}
