% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ALL_STRATA.R
\name{make_all_stratum}
\alias{make_all_stratum}
\title{Make All Stratum}
\usage{
make_all_stratum(fn022, fn023, fn024, fn026, fn028)
}
\arguments{
\item{fn022}{FN022 design table}

\item{fn023}{FN023 design table}

\item{fn024}{FN024 design table}

\item{fn026}{FN026 design table}

\item{fn028}{FN028 design table}
}
\value{
dataframe with all possible strata
}
\description{
Make All Stratum
}
\examples{
\dontrun{
make_all_stratum(SC00$FN022, SC00$FN023, SC00$FN024, SC00$FN026, SC00$FN028)
}


}
