% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FR712.R
\name{make_FR712}
\alias{make_FR712}
\title{make_FR712}
\usage{
make_FR712(fn022 = FN022, fn023 = FN023, fn025 = FN025)
}
\arguments{
\item{fn025}{}
}
\value{
FR712 table
}
\description{
This function takes FN tables as inputs and generates the FR712 table.
}
