% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FR713714.R
\name{make_FR713714}
\alias{make_FR713714}
\title{make_FR713714}
\usage{
make_FR713714(
  fn011,
  fn022,
  fn023,
  fn024,
  fn025,
  fn026,
  fn028,
  fn111,
  fn112,
  fn121,
  fn123,
  fileloc
)
}
\arguments{
\item{fn011}{FN project definition table FN011 which lists project dates, year, lake name and project leader information}

\item{fn022}{FN strata definition file for SSN}

\item{fn023}{FN strata definition file for DTP i.e. workday and weekend}

\item{fn024}{FN stratum definition for period i.e. am or pm}

\item{fn025}{FN stratum definition listing the exception days i.e. holidays that need to be re-coded as weekends}

\item{fn026}{FN stratum definition for the SPACE}

\item{fn028}{FN stratum definition for the MODE}

\item{fn111}{FN portion of the CLF that outlines the SAMA, Date, DOW, STRATUM,}

\item{fn112}{FN portion of the CLF that outlines the activity count, time of the count and number of interviews that occurred. Note: multiple counts could occur per SAMA}

\item{fn121}{FN CIF creel interview form}

\item{fn123}{FN species records part of the CIF form but nested by SAM to the FN121}

\item{fileloc}{FN stratum definition for the MODE}
}
\description{
This function generates the estimates and outputs a table for EFFORT and HARVEST
}
