% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FR713_table.R
\name{make_FR713_table}
\alias{make_FR713_table}
\title{make_FR713_table}
\usage{
make_FR713_table(fr713)
}
\arguments{
\item{fr713}{the output of the make_FR713 function}

\item{fr714}{the output of the make_FR714 function}

\item{Table}{the effort table will print if yes}

\item{EffortHarvest}{both the effort and harvest table will print if yes}
}
\value{
a table of the creel estimates
}
\description{
This is the creel estimate table for EFFORT
}
