% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_STRAT.R
\name{parse_STRAT}
\alias{parse_STRAT}
\title{parse_STRAT}
\usage{
parse_STRAT(fndf)
}
\arguments{
\item{fndf}{a FN2 data table that includes STRAT as a column}
}
\value{
return the same FN2 dataframe with SSN, DTP, PRD, SPACE, and MODE columns as factors
}
\description{
STRAT is an important grouping variable used throughout FN2. Often though
it is necessary to have the individual components parsed out in to data columns
in order to be able to join tables together.
}
\examples{
dummydf <- data.frame(STRAT = c("01_13_01_02", "01_11_02_01", "03_22_04_01"))
parse_STRAT(dummydf)
}
