% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC00.R
\name{SC00}
\alias{SC00}
\title{SC00_XXX example data}
\format{
list with each FN table as a dataframe in the list
}
\value{
list of data tables
}
\description{
This is example data included in the package for testing and vignettes.
}
\examples{
\dontrun{
data("SC00")
names(SC00)
lapply(SC00, head)
SC00$FN011
}

}
