% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FR713.R
\name{make_FR713}
\alias{make_FR713}
\title{Make FR713}
\usage{
make_FR713(fn022, fn023, fn024, fn025, fn026, fn028, fn111, fn112, fn121)
}
\arguments{
\item{fn121}{}
}
\value{
FR713-like table
}
\description{
Make FR713
}
\examples{
\dontrun{
make_FR713(SC00$FN022, SC00$FN023, SC00$FN024, SC00$FN025, SC00$FN026, SC00$FN028, SC00$FN112, SC00$FN121)
}
}
