% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_FR713.R
\name{make_FR713}
\alias{make_FR713}
\title{make_FR713}
\usage{
make_FR713(fn011, fn111, fn112, fn121, fn024)
}
\arguments{
\item{fn011}{FN project definition table FN011 which lists project dates, year, lake name and project leader information}

\item{fn111}{FN portion of the CLF that oulines the SAMA, Date, DOW, STRATUM,}

\item{fn112}{FN portion of the CLF that outlines the activity count, time of the count and number of interviews that occured. Note: multiple counts could occur per SAMA}

\item{fn121}{FN CIF creel interview form}

\item{fn024}{FN stratum definition for period i.e. am or pm}
}
\description{
This is the creel estimate table for EFFORT
}
